% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{hi_there}
\alias{hi_there}
\title{A friendly greeting to a friendly person.}
\usage{
hi_there(name)
}
\arguments{
\item{name}{A name}
}
\value{
Hi there name!
}
\description{
A friendly greeting to a friendly person.
}
\examples{

hi_there("Tim")
}
